(refactor) move htmlxtojsx into own folder
