+++ packages/svelte2tsx/src/htmlxtojsx/nodes/component-type.ts
