+++ packages/svelte2tsx/src/htmlxtojsx/svgattributes.ts
